// src/app/generated/gtfs-realtime.ts
export interface TripUpdate_StopTimeEvent {
  /**
   * Delay (in seconds) can be positive (meaning that the vehicle is late) or
   * negative (meaning that the vehicle is ahead of schedule). Delay of 0
   * means that the vehicle is exactly on time.
   */
  delay?:
    | number
    | undefined;
  /**
   * Event as absolute time.
   * In Unix time (i.e., number of seconds since January 1st 1970 00:00:00
   * UTC).
   */
  time?:
    | number
    | undefined;
  /**
   * If uncertainty is omitted, it is interpreted as unknown.
   * If the prediction is unknown or too uncertain, the delay (or time) field
   * should be empty. In such case, the uncertainty field is ignored.
   * To specify a completely certain prediction, set its uncertainty to 0.
   */
  uncertainty?:
    | number
    | undefined;
  /**
   * Scheduled time for a NEW, REPLACEMENT, or DUPLICATED trip.
   * In Unix time (i.e., number of seconds since January 1st 1970 00:00:00
   * UTC).
   * Optional if TripUpdate.schedule_relationship is NEW, REPLACEMENT or DUPLICATED, forbidden otherwise.
   * NOTE: This field is still experimental, and subject to change. It may be formally adopted in the future.
   */
  scheduledTime?: number | undefined;
}

/**
 * Realtime update for arrival and/or departure events for a given stop on a
 * trip. Updates can be supplied for both past and future events.
 * The producer is allowed, although not required, to drop past events.
 */
export interface TripUpdate_StopTimeUpdate {
  /** Must be the same as in stop_times.txt in the corresponding GTFS feed. */
  stopSequence?:
    | number
    | undefined;
  /** Must be the same as in stops.txt in the corresponding GTFS feed. */
  stopId?: string | undefined;
  arrival?: TripUpdate_StopTimeEvent | undefined;
  departure?:
    | TripUpdate_StopTimeEvent
    | undefined;
  /**
   * Expected occupancy after departure from the given stop.
   * Should be provided only for future stops.
   * In order to provide departure_occupancy_status without either arrival or
   * departure StopTimeEvents, ScheduleRelationship should be set to NO_DATA.
   */
  departureOccupancyStatus?: VehiclePosition_OccupancyStatus | undefined;
  scheduleRelationship?:
    | TripUpdate_StopTimeUpdate_ScheduleRelationship
    | undefined;
  /**
   * Realtime updates for certain properties defined within GTFS stop_times.txt
   * NOTE: This field is still experimental, and subject to change. It may be formally adopted in the future.
   */
  stopTimeProperties?: TripUpdate_StopTimeUpdate_StopTimeProperties | undefined;
}

/** The relation between the StopTimeEvents and the static schedule. */
export enum TripUpdate_StopTimeUpdate_ScheduleRelationship {
  /**
   * SCHEDULED - The vehicle is proceeding in accordance with its static schedule of
   * stops, although not necessarily according to the times of the schedule.
   * At least one of arrival and departure must be provided. If the schedule
   * for this stop contains both arrival and departure times then so must
   * this update. Frequency-based trips (GTFS frequencies.txt with exact_times = 0)
   * should not have a SCHEDULED value and should use UNSCHEDULED instead.
   */
  SCHEDULED = 0,
  /**
   * SKIPPED - The stop is skipped, i.e., the vehicle will not stop at this stop.
   * Arrival and departure are optional.
   */
  SKIPPED = 1,
  /**
   * NO_DATA - No StopTimeEvents are given for this stop.
   * The main intention for this value is to give time predictions only for
   * part of a trip, i.e., if the last update for a trip has a NO_DATA
   * specifier, then StopTimeEvents for the rest of the stops in the trip
   * are considered to be unspecified as well.
   * Neither arrival nor departure should be supplied.
   */
  NO_DATA = 2,
  /**
   * UNSCHEDULED - The vehicle is operating a trip defined in GTFS frequencies.txt with exact_times = 0.
   * This value should not be used for trips that are not defined in GTFS frequencies.txt,
   * or trips in GTFS frequencies.txt with exact_times = 1. Trips containing StopTimeUpdates
   * with ScheduleRelationship=UNSCHEDULED must also set TripDescriptor.ScheduleRelationship=UNSCHEDULED.
   * NOTE: This field is still experimental, and subject to change. It may be
   * formally adopted in the future.
   */
  UNSCHEDULED = 3,
  UNRECOGNIZED = -1,
}
// ...
export interface TripUpdate_StopTimeUpdate_StopTimeProperties {
  /**
   * Supports real-time stop assignments. Refers to a stop_id defined in the GTFS stops.txt.
   * The new assigned_stop_id should not result in a significantly different trip experience for the end user than
   * the stop_id defined in GTFS stop_times.txt. In other words, the end user should not view this new stop_id as an
   * "unusual change" if the new stop was presented within an app without any additional context.
   * For example, this field is intended to be used for platform assignments by using a stop_id that belongs to the
   * same station as the stop originally defined in GTFS stop_times.txt.
   * To assign a stop without providing any real-time arrival or departure predictions, populate this field and set
   * StopTimeUpdate.schedule_relationship = NO_DATA.
   * If this field is populated, it is preferred to omit `StopTimeUpdate.stop_id` and use only `StopTimeUpdate.stop_sequence`. If
   * `StopTimeProperties.assigned_stop_id` and `StopTimeUpdate.stop_id` are populated, `StopTimeUpdate.stop_id` must match `assigned_stop_id`.
   * Platform assignments should be reflected in other GTFS-realtime fields as well
   * (e.g., `VehiclePosition.stop_id`).
   * NOTE: This field is still experimental, and subject to change. It may be formally adopted in the future.
   */
  assignedStopId?:
    | string
    | undefined;
  /**
   * The updated headsign of the vehicle at the stop.
   * NOTE: This field is still experimental, and subject to change. It may be formally adopted in the future.
   */
  stopHeadsign?:
    | string
    | undefined;
  /**
   * The updated pickup of the vehicle at the stop.
   * NOTE: This field is still experimental, and subject to change. It may be formally adopted in the future.
   */
  pickupType?:
    | TripUpdate_StopTimeUpdate_StopTimeProperties_DropOffPickupType
    | undefined;
  /**
   * The updated drop off of the vehicle at the stop.
   * NOTE: This field is still experimental, and subject to change. It may be formally adopted in the future.
   */
  dropOffType?: TripUpdate_StopTimeUpdate_StopTimeProperties_DropOffPickupType | undefined;
}
// ...
export enum VehiclePosition_OccupancyStatus {
  /**
   * EMPTY - The vehicle or carriage is considered empty by most measures, and has few or no
   * passengers onboard, but is still accepting passengers.
   */
  EMPTY = 0,
  /**
   * MANY_SEATS_AVAILABLE - The vehicle or carriage has a large number of seats available.
   * The amount of free seats out of the total seats available to be
   * considered large enough to fall into this category is determined at the
   * discretion of the producer.
   */
  MANY_SEATS_AVAILABLE = 1,
  /**
   * FEW_SEATS_AVAILABLE - The vehicle or carriage has a relatively small number of seats available.
   * The amount of free seats out of the total seats available to be
   * considered small enough to fall into this category is determined at the
   * discretion of the feed producer.
   */
  FEW_SEATS_AVAILABLE = 2,
  /** STANDING_ROOM_ONLY - The vehicle or carriage can currently accommodate only standing passengers. */
  STANDING_ROOM_ONLY = 3,
  /**
   * CRUSHED_STANDING_ROOM_ONLY - The vehicle or carriage can currently accommodate only standing passengers
   * and has limited space for them.
   */
  CRUSHED_STANDING_ROOM_ONLY = 4,
  /**
   * FULL - The vehicle or carriage is considered full by most measures, but may still be
   * allowing passengers to board.
   */
  FULL = 5,
  /** NOT_ACCEPTING_PASSENGERS - The vehicle or carriage is not accepting passengers, but usually accepts passengers for boarding. */
  NOT_ACCEPTING_PASSENGERS = 6,
  /** NO_DATA_AVAILABLE - The vehicle or carriage doesn't have any occupancy data available at that time. */
  NO_DATA_AVAILABLE = 7,
  /**
   * NOT_BOARDABLE - The vehicle or carriage is not boardable and never accepts passengers.
   * Useful for special vehicles or carriages (engine, maintenance carriage, etc…).
   */
  NOT_BOARDABLE = 8,
  UNRECOGNIZED = -1,
}
// ...
export const TripUpdate_StopTimeEvent: MessageFns<TripUpdate_StopTimeEvent> = {
  encode(message: TripUpdate_StopTimeEvent, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.delay !== undefined && message.delay !== 0) {
      writer.uint32(8).int32(message.delay);
    }
    if (message.time !== undefined && message.time !== 0) {
      writer.uint32(16).int64(message.time);
    }
    if (message.uncertainty !== undefined && message.uncertainty !== 0) {
      writer.uint32(24).int32(message.uncertainty);
    }
    if (message.scheduledTime !== undefined && message.scheduledTime !== 0) {
      writer.uint32(32).int64(message.scheduledTime);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): TripUpdate_StopTimeEvent {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseTripUpdate_StopTimeEvent();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 8) {
            break;
          }

          message.delay = reader.int32();
          continue;
        }
        case 2: {
          if (tag !== 16) {
            break;
          }

          message.time = longToNumber(reader.int64());
          continue;
        }
        case 3: {
          if (tag !== 24) {
            break;
          }

          message.uncertainty = reader.int32();
          continue;
        }
        case 4: {
          if (tag !== 32) {
            break;
          }

          message.scheduledTime = longToNumber(reader.int64());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): TripUpdate_StopTimeEvent {
    return {
      delay: isSet(object.delay) ? globalThis.Number(object.delay) : 0,
      time: isSet(object.time) ? globalThis.Number(object.time) : 0,
      uncertainty: isSet(object.uncertainty) ? globalThis.Number(object.uncertainty) : 0,
      scheduledTime: isSet(object.scheduledTime) ? globalThis.Number(object.scheduledTime) : 0,
    };
  },

  toJSON(message: TripUpdate_StopTimeEvent): unknown {
    const obj: any = {};
    if (message.delay !== undefined && message.delay !== 0) {
      obj.delay = Math.round(message.delay);
    }
    if (message.time !== undefined && message.time !== 0) {
      obj.time = Math.round(message.time);
    }
    if (message.uncertainty !== undefined && message.uncertainty !== 0) {
      obj.uncertainty = Math.round(message.uncertainty);
    }
    if (message.scheduledTime !== undefined && message.scheduledTime !== 0) {
      obj.scheduledTime = Math.round(message.scheduledTime);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<TripUpdate_StopTimeEvent>, I>>(base?: I): TripUpdate_StopTimeEvent {
    return TripUpdate_StopTimeEvent.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<TripUpdate_StopTimeEvent>, I>>(object: I): TripUpdate_StopTimeEvent {
    const message = createBaseTripUpdate_StopTimeEvent();
    message.delay = object.delay ?? 0;
    message.time = object.time ?? 0;
    message.uncertainty = object.uncertainty ?? 0;
    message.scheduledTime = object.scheduledTime ?? 0;
    return message;
  },
};
// ...
export interface MessageFns<T> {
  encode(message: T, writer?: BinaryWriter): BinaryWriter;
  decode(input: BinaryReader | Uint8Array, length?: number): T;
  fromJSON(object: any): T;
  toJSON(message: T): unknown;
  create<I extends Exact<DeepPartial<T>, I>>(base?: I): T;
  fromPartial<I extends Exact<DeepPartial<T>, I>>(object: I): T;
}